+++ catalog/src/components/ui/index.js
@@ -1,9 +1,11 @@
 import ButtomSrc from "./ButtomSrc";
 import CardPost from "./CardPost";
 import Preloader from "./Preloader";
+import Chits from "./Chits";
 
 export default [
     ButtomSrc,
     CardPost,
     Preloader,
+    Chits,
 ]
